+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # ChangeLog
 
 - bump Ruby 2.6.9, 2.7.5, 3.0.3 in CI.
-- check duckdb > 0.1.1 in extconf.rb.
+- check duckdb > 0.2.5 in extconf.rb.
+  - breaking change. ruby-duckdb requires duckdb 0.2.5 or later.
 
 # 0.3.1.0